requirements.txt
certifi==2020.11.8
cycler==0.10.0
dataclasses==0.6
freetype-py==2.2.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.3.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-msssim==0.2.1
pytz==2020.4
six==1.15.0
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
